more organized testing for the catheter table
